requirements.txt
whois
dnspython
cryptography
tkinter
reportlab
